requirements.txt
fastapi>=0.109.2
starlette>=0.36.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary
bcrypt==3.2.2
passlib[bcrypt]
python-jose[cryptography]
pydantic>=2.0.0
python-multipart
jinja2>=3.1.0
paho-mqtt>=1.6.1